Copy wavepacket objects when setting up sources
Every time a source is setup, its configurations change.  However, due
to Python's copy by reference, if two or sources are the same instance
but with different positions, its configuration will be that of the
last position.  To avoid all sources having wrong configurations, it
is necessary to copy the deep copy the objects.