Change calling convention of setter methods
The call for the 'set_{var_name}' method after being found by getattr
should take only one argument.  This restriction implies that the 'set_'
methods should be defined inside a class definition, before the object
is instantiated.  This may confuse newcomers, but it is the usual
approach with Python.  In other words, to user should not do this:

    obj = SuperClass()
    obj.foo = <value>
    def set_foo(self, value):
        ...
    obj.set_foo = set_foo

and do this instead:

    class ExtendedClass(SuperClass):
    	  foo = <value>
    	  def set_foo(self, value):
	      ...
    obj = ExtendedClass()

The drawback of this convention is that now there will class variables
shared across instances, but this module is designed to have only one
single instance of each class -- and probably from only one class ---
at any given time.  The consequences of this decisions remain to be
seen.