Add SingleDevice.__del__
Overload `__del__` to catch `AttributeError` exceptions.

When using `with SingleDevice(<devname>)` and `<devname>` is not
valid, the read and write tasks are not created and an exception is
raised.  However, before raising it, the interpreter leaves the `with`
block by calling `__del__`.  If there is nothing to be deleted, a
second exception will be raised.  To keep only the exceptions related
to the invalid device, catch exceptions when trying to delete the
tasks that were not created.