Take initial values
Sometimes I needed to create a dataclass that ended up being saved
inside MeasuredData.  When trying to load the object, I had to
redefine the class to be able to unpickle it.  Instead o saving a
dataclass, it is better to just create a MeasuredData with some
initial values.  It is less clear, and type hints cannot be used, but
it saves the hassle of defining classes to be able to load objects.